EstimateSwapExactAmountIn to osmocli  and swaprouter
