+++ x/gamm/client/cli/query.go
@@ -26,7 +26,6 @@ func GetQueryCmd() *cobra.Command {
 	osmocli.AddQueryCmd(cmd, types.NewQueryClient, GetCmdSpotPrice)
 	osmocli.AddQueryCmd(cmd, types.NewQueryClient, GetCmdPool)
 	osmocli.AddQueryCmd(cmd, types.NewQueryClient, GetCmdPools)
-	osmocli.AddQueryCmd(cmd, types.NewQueryClient, GetCmdEstimateSwapExactAmountIn)
 	osmocli.AddQueryCmd(cmd, types.NewQueryClient, GetCmdEstimateSwapExactAmountOut)
 	cmd.AddCommand(
 		GetCmdPoolParams(),
@@ -184,18 +183,6 @@ func GetCmdSpotPrice() (*osmocli.QueryDescriptor, *types.QuerySpotPriceRequest) 
 `}, &types.QuerySpotPriceRequest{}
 }
 
-// GetCmdEstimateSwapExactAmountIn returns estimation of output coin when amount of x token input.
-func GetCmdEstimateSwapExactAmountIn() (*osmocli.QueryDescriptor, *types.QuerySwapExactAmountInRequest) {
-	return &osmocli.QueryDescriptor{
-		Use:   "estimate-swap-exact-amount-in <poolID> <sender> <tokenIn>",
-		Short: "Query estimate-swap-exact-amount-in",
-		Long: `Query estimate-swap-exact-amount-in.{{.ExampleHeader}}
-{{.CommandPrefix}} estimate-swap-exact-amount-in 1 osm11vmx8jtggpd9u7qr0t8vxclycz85u925sazglr7 stake --swap-route-pool-ids=2 --swap-route-pool-ids=3`,
-		ParseQuery: EstimateSwapExactAmountInParseArgs,
-		Flags:      osmocli.FlagDesc{RequiredFlags: []*flag.FlagSet{FlagSetMultihopSwapRoutes()}},
-	}, &types.QuerySwapExactAmountInRequest{}
-}
-
 // GetCmdEstimateSwapExactAmountOut returns estimation of input coin to get exact amount of x token output.
 func GetCmdEstimateSwapExactAmountOut() (*osmocli.QueryDescriptor, *types.QuerySwapExactAmountOutRequest) {
 	return &osmocli.QueryDescriptor{
@@ -206,25 +193,6 @@ func GetCmdEstimateSwapExactAmountOut() (*osmocli.QueryDescriptor, *types.QueryS
 		ParseQuery: EstimateSwapExactAmountOutParseArgs,
 		Flags:      osmocli.FlagDesc{RequiredFlags: []*flag.FlagSet{FlagSetMultihopSwapRoutes()}},
 	}, &types.QuerySwapExactAmountOutRequest{}
-}
-
-func EstimateSwapExactAmountInParseArgs(args []string, fs *flag.FlagSet) (proto.Message, error) {
-	poolID, err := strconv.Atoi(args[0])
-	if err != nil {
-		return nil, err
-	}
-
-	routes, err := swapAmountInRoutes(fs)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.QuerySwapExactAmountInRequest{
-		Sender:  args[1],        // TODO: where sender is used?
-		PoolId:  uint64(poolID), // TODO: is this poolId used?
-		TokenIn: args[2],
-		Routes:  routes,
-	}, nil
 }
 
 func EstimateSwapExactAmountOutParseArgs(args []string, fs *flag.FlagSet) (proto.Message, error) {
