+++ x/gamm/client/cli/tx.go
@@ -356,37 +356,6 @@ func stringArrayCoinsParser(flagName string, fs *flag.FlagSet) (sdk.Coins, error
 	return coins, nil
 }
 
-func swapAmountInRoutes(fs *flag.FlagSet) ([]types.SwapAmountInRoute, error) {
-	swapRoutePoolIds, err := fs.GetString(FlagSwapRoutePoolIds)
-	swapRoutePoolIdsArray := strings.Split(swapRoutePoolIds, ",")
-	if err != nil {
-		return nil, err
-	}
-
-	swapRouteDenoms, err := fs.GetString(FlagSwapRouteDenoms)
-	swapRouteDenomsArray := strings.Split(swapRouteDenoms, ",")
-	if err != nil {
-		return nil, err
-	}
-
-	if len(swapRoutePoolIdsArray) != len(swapRouteDenomsArray) {
-		return nil, errors.New("swap route pool ids and denoms mismatch")
-	}
-
-	routes := []types.SwapAmountInRoute{}
-	for index, poolIDStr := range swapRoutePoolIdsArray {
-		pID, err := strconv.Atoi(poolIDStr)
-		if err != nil {
-			return nil, err
-		}
-		routes = append(routes, types.SwapAmountInRoute{
-			PoolId:        uint64(pID),
-			TokenOutDenom: swapRouteDenomsArray[index],
-		})
-	}
-	return routes, nil
-}
-
 func swapAmountOutRoutes(fs *flag.FlagSet) ([]types.SwapAmountOutRoute, error) {
 	swapRoutePoolIds, err := fs.GetString(FlagSwapRoutePoolIds)
 	swapRoutePoolIdsArray := strings.Split(swapRoutePoolIds, ",")
