+++ x/swaprouter/client/cli/common.go
@@ -11,29 +11,31 @@ import (
 )
 
 func swapAmountInRoutes(fs *flag.FlagSet) ([]types.SwapAmountInRoute, error) {
-	swapRoutePoolIds, err := fs.GetStringArray(FlagSwapRoutePoolIds)
+	swapRoutePoolIds, err := fs.GetString(FlagSwapRoutePoolIds)
+	swapRoutePoolIdsArray := strings.Split(swapRoutePoolIds, ",")
 	if err != nil {
 		return nil, err
 	}
 
-	swapRouteDenoms, err := fs.GetStringArray(FlagSwapRouteDenoms)
+	swapRouteDenoms, err := fs.GetString(FlagSwapRouteDenoms)
+	swapRouteDenomsArray := strings.Split(swapRouteDenoms, ",")
 	if err != nil {
 		return nil, err
 	}
 
-	if len(swapRoutePoolIds) != len(swapRouteDenoms) {
+	if len(swapRoutePoolIdsArray) != len(swapRouteDenomsArray) {
 		return nil, errors.New("swap route pool ids and denoms mismatch")
 	}
 
 	routes := []types.SwapAmountInRoute{}
-	for index, poolIDStr := range swapRoutePoolIds {
+	for index, poolIDStr := range swapRoutePoolIdsArray {
 		pID, err := strconv.Atoi(poolIDStr)
 		if err != nil {
 			return nil, err
 		}
 		routes = append(routes, types.SwapAmountInRoute{
 			PoolId:        uint64(pID),
-			TokenOutDenom: swapRouteDenoms[index],
+			TokenOutDenom: swapRouteDenomsArray[index],
 		})
 	}
 	return routes, nil
